requirements/requirements.txt
Django==1.10.2
psycopg2==2.6.2
Pygments==2.1.3
django-easy-select2==1.3.3
djangorestframework==3.5.2
django-filter==0.15.3